+++ src/routes/tainted/+page.svelte
@@ -32,7 +32,8 @@
 
 <SuperDebug data={{ $tainted }} />
 
-<a href="/tainted/multiple-tainted">Multiple tainted &gt;</a>
+<a href="/tainted/multiple-tainted">Multiple tainted &gt;</a> |
+<a href="/tainted/programmatically">Programmatically &gt;</a>
 
 <h2>Tainted modification test</h2>